Fixed timer 0 and 2 conflicts by reducing frequency of interrupts

+++ include/timing/timing.h
@@ -6,7 +6,7 @@
 void timer0_init_as_pwm(void);
 void timer0_set_duty_cycle_a(float duty_cycle);
 void timer0_set_duty_cycle_b(float duty_cycle);
-void timer0_init_as_timer_ms(uint32_t ms, void (*callback)(void));
+void timer0_init_as_timer_ms(float ms, void (*callback)(void));
 
 void timer1_init_as_pwm(void);
 void timer1_set_duty_cycle_a(float duty_cycle);
@@ -16,6 +16,6 @@ void timer1_init_as_timer_ms(uint32_t ms, void (*callback)(void));
 void timer2_init_as_pwm(void);
 void timer2_set_duty_cycle_a(float duty_cycle);
 void timer2_set_duty_cycle_b(float duty_cycle);
-void timer2_init_as_timer_ms(uint32_t ms, void (*callback)(void));
+void timer2_init_as_timer_ms(float ms, void (*callback)(void));
 
 #endif
